cross compile to scalaJS
