+++ jts/src/main/com/monsanto/labs/mwundo/Java2Dable.java
@@ -1,0 +1,8 @@
+package com.monsanto.labs.mwundo;
+
+/**
+ * Created by Ryan Richt on 10/26/15
+ */
+trait Java2Dable[G] {
+  def toJava2D(g: G): Seq[java.awt.Shape]
+}
